* Made many modifications to the engine  * Converted BasePoint from Int to Float.
git-svn-id: http://subversion.assembla.com/svn/Project_Valkyrie2@178 080cfd25-d5d9-4c59-82ea-70d02a200d8b

+++ trunk/src/ValkyrieEngine/Core/Points/MapPoint.cs
@@ -12,7 +12,7 @@ namespace Valkyrie.Engine.Core
     // one unit in a map point represents 32 units in a screen point
     public class MapPoint : BasePoint
     {
-        public MapPoint(int x, int y) : base(x,y)
+        public MapPoint(float x, float y) : base(x, y)
         {
         }
